Balancing, fix bug 1

+++ AI_P01/AI_P01/Weapon.h
@@ -30,13 +30,13 @@ struct FWeapon
     struct Data
     {
         Data()
-            : weapon(EWeapon::SMG), weaponType(EWeaponType::Gun), damage(20), actionsPerSecond(8.0f), hasAmmo(true), ammo(100), level(1), name("Weapon") {}
+            : weapon(EWeapon::SMG), weaponType(EWeaponType::Gun), damage(20.0f), actionsPerSecond(8.0f), hasAmmo(true), ammo(100), level(1), name("Weapon") {}
         Data(EWeapon _weapon, EWeaponType _weaponType, int _damage, float _actionsPerSecond, int _ammo, float _bulletSpeed, const char* _name)
             : weapon(_weapon), weaponType(_weaponType), damage(_damage), actionsPerSecond(_actionsPerSecond), hasAmmo(true), ammo(_ammo), speed(_bulletSpeed), level(1), name(_name) {}
 
         EWeapon weapon;
         EWeaponType weaponType;
-        int damage;
+        float damage;
         float actionsPerSecond;
         float speed;
         bool hasAmmo;
